Config changes

+++ YOLOv3_Train_Inference/deploy.py
@@ -27,13 +27,64 @@ def service_exists(namespace, service_name):
             return True
     return False
 
-def generate_yamls(template_yaml_path, config_path):
-    yamls = []
+def generate_yamls(path, template_yaml_path, config_path, namespace):
+    server_yamls = []
+    client_yamls = []
+    network_yamls = []
     config = json.load(open(config_path))
-    print(config)
-    return yamls
+    for machine in config['machines']:
+        # load basic template
+        with open(template_yaml_path, 'r') as f:
+            template = yaml.safe_load(f)
+
+        # write deployment yaml content
+        template['metadata']['name'] = machine['name']
+        template['metadata']['namespace'] = namespace
+        template['metadata']['labels'] = {'app':machine['name']}
+        template['spec']['selector']['matchLabels'] = {'app':machine['name']}
+        template['spec']['template']['metadata']['labels'] = {'app':machine['name']}
+        template['spec']['template']['spec']['containers'][0]['name'] = machine['name']
+        template['spec']['template']['spec']['containers'][0]['image'] = machine['image']
+        template['spec']['template']['spec']['containers'][0]['command'] = machine['command']
+        template['spec']['template']['spec']['containers'][0]['args'] = machine['args']
+        template['spec']['template']['spec']['containers'][0]['resources'] = machine['resources']
+        template['spec']['template']['spec']['containers'][0]['volumeMounts'] = []
+        template['spec']['template']['spec']['containers'][0]['ports'] = [
+            {"containerPort": int(machine['port'])}
+        ]
+        template['spec']['template']['spec']['volumes'] = []
+        for volume in machine['volumeMounts']:
+            template['spec']['template']['spec']['containers'][0]['volumeMounts'].append(
+                {"name": volume['name'], "mountPath": volume['mountPath']}
+            )
+            template['spec']['template']['spec']['volumes'].append(
+                {"name": volume['name'], "nfs": volume['nfs']}
+            )
+
+        # write service yaml content
+        service_template = {'apiVersion': 'v1', 'kind': 'Service', 'metadata': {}, 'spec': {}}
+        service_template['metadata']['name'] = machine['name'] + '-service'
+        service_template['metadata']['namespace'] = namespace
+        service_template['spec']['selector'] = {'app':machine['name']}
+        if machine['category'] == 'server':
+            service_template['spec']['ports'] = [{'protocol': 'TCP', 'port': 9876}]
+        elif machine['category'] == 'client':
+            service_template['spec']['ports'] = [{'protocol': 'TCP', 'port': 8080, 'targetPort': 9876}]
+
+        yaml_path = path + '/' + machine['name'] + '.yaml'
+        with open(yaml_path, 'w') as f:
+            yaml.safe_dump_all([template, service_template], f)
+
+        if machine['category'] == 'server':
+            server_yamls.append(yaml_path)
+        elif machine['category'] == 'client':
+            client_yamls.append(yaml_path)
+
+
+    return server_yamls, client_yamls, network_yamls
 
 def launch_server(server_yaml_path, namespace, core_client, app_client):
+    print("Launching server: ", server_yaml_path)
     server_info = {}
 
     with open(server_yaml_path, 'r') as f:
@@ -78,30 +129,31 @@ def launch_server(server_yaml_path, namespace, core_client, app_client):
     return server_info
 
 def launch_client(server_info, client_yaml_path, 
-                    namespace, core_client, app_client, client_id):
+                    namespace, core_client, app_client):
+    print("Launching client: ", client_yaml_path)
+
     client_info = {}
 
     with open(client_yaml_path, 'r') as f:
         deps = yaml.safe_load_all(f)
         try:
             for dep in deps:
-                dep['metadata']['name'] = dep['metadata']['name'].format(client_id)
                 name = dep['metadata']['name']
                 if dep['kind'] == 'Deployment':
                     client_info['name'] = name
                     client_info['label'] = dep['metadata']['labels']['app']
                     client_info['existing_pods'] = []
                     dep['spec']['template']['spec']['containers'][0]['args'][1] = \
-                          dep['spec']['template']['spec']['containers'][0]['args'][1].format(server_info['IP'], client_id)
+                          dep['spec']['template']['spec']['containers'][0]['args'][1].format(server_info['IP'])
                     for volume in dep['spec']['template']['spec']['volumes']:
-                        volume['nfs']['path'] = volume['nfs']['path'] + "/" + str(client_id)
                         user_path = re.sub('/users/', '/h/', volume['nfs']['path'])
-                        pathlib.Path(user_path).mkdir(parents=True, exist_ok=True)
-                        os.chmod(user_path, 0o777)
-                        if volume['name'] == 'dataset-volume':
-                            if not os.listdir(user_path):
-                                shutil.copy2('/h/churongj/ACAI_Emulator/YOLOv3_Train_Inference/content/VOCtrainval_06-Nov-2007.tar', 
-                                                user_path)
+                        if not os.path.exists(user_path):
+                            pathlib.Path(user_path).mkdir(parents=True, exist_ok=True)
+                            os.chmod(user_path, 0o777)
+                            if volume['name'] == 'dataset-volume':
+                                if not os.listdir(user_path):
+                                    shutil.copy2('/h/churongj/ACAI_Emulator/YOLOv3_Train_Inference/content/VOCtrainval_06-Nov-2007.tar', 
+                                                    user_path)
                     if deployment_exists(namespace, name):
                         client_pods = core_client.list_namespaced_pod(namespace, label_selector="app=" + client_info['label']).items
                         for pod in client_pods:
@@ -132,23 +184,23 @@ def launch_client(server_info, client_yaml_path,
 
     return client_info
 
-def main(path, client_num):
+def main(path):
     config.load_kube_config()
 
     server_yaml_path = 'serverconfig.yaml'
     client_yaml_path = 'clientconfig.yaml'
-    namespace = 'yolo'
+    namespace = 'yolo-test'
     app_client = client.AppsV1Api()
     core_client = client.CoreV1Api()
     configuration = client.Configuration()
 
     template_yaml_path = 'yaml_templates/config.yaml'
     config_path = 'user_config.json'
-    yamls = generate_yamls(template_yaml_path, config_path)
+    server_yamls, client_yamls, network_yamls = generate_yamls(path, template_yaml_path, config_path, namespace)
 
-    server_info = launch_server(server_yaml_path, namespace, core_client, app_client)
-    for i in range(client_num):
-        client_info = launch_client(server_info, client_yaml_path, namespace, core_client, app_client, i)
+    server_info = launch_server(server_yamls[0], namespace, core_client, app_client)
+    for client_yaml_path in client_yamls:
+        client_info = launch_client(server_info, client_yaml_path, namespace, core_client, app_client)
 
     time.sleep(20)
 
@@ -187,5 +239,4 @@ def main(path, client_num):
 
 if __name__ == '__main__':
     path = sys.argv[1]
-    client_num = int(sys.argv[2])
-    main(path, client_num)
+    main(path)
